==> 04_predict_stress.ipynb <==
from openvino.runtime import Core
import numpy as np
import pandas as pd

# Load Model
ie = Core()
model_path = "./models/battery_stress_model/saved_model.xml"
compiled_model = ie.compile_model(model_path, "CPU")
input_layer = compiled_model.input(0)
output_layer = compiled_model.output(0)

# New Data
new_data = pd.DataFrame([{
    "stateOfCharge": 0.15,  # Low SoC
    "stateOfHealth": 97.00,
    "batteryCurrent": 480.00,  # High current
    "batteryVoltage": 355.00,  # Low voltage
    "kmh": 150.00,
    "distance": 50.00,
    "batteryTemp": 55.00,  # High temp
    "ambientTemp": 22.00
}])

input_data = new_data.astype(np.float32).to_numpy()

# Make Prediction
output = compiled_model([input_data])
predicted_class = np.argmax(output[output_layer], axis=1)
print("Stress Prediction:", "STRESSED" if output[0] == 1 else "NORMAL")
