==> 06_predict_ttf.ipynb <==
from openvino.runtime import Core
import numpy as np
import pandas as pd

# Load Model
ie = Core()
model_path = "./models/battery_ttf_model/saved_model.xml"
compiled_model = ie.compile_model(model_path, "CPU")
input_layer = compiled_model.input(0)
output_layer = compiled_model.output(0)

# New Battery Data
input_data = np.array([[45.0, 500.0, 370.0, 0.4, 98.0]], dtype=np.float32)  # [Temp, Current, Voltage, SoC, SoH]

# Predict Time Before Failure
output = compiled_model([input_data])
ttf_prediction = output[output_layer]
predicted_ttf = ttf_prediction[0][0].item()
print(f"Predicted Time Before Failure: {predicted_ttf:.2f} hours")
