==> 07_save_models.ipynb <==
# Save model

In this notebook we are going to create an s3 connection to upload our model

# Imports for the s3 connection
# ==============================================================================
import os
import boto3

# Read the data connection variables
# ==============================================================================
key_id = os.getenv("AWS_ACCESS_KEY_ID")
secret_key = os.getenv("AWS_SECRET_ACCESS_KEY")
endpoint = os.getenv("AWS_S3_ENDPOINT")
bucket_name = os.getenv("AWS_S3_BUCKET")

# Create s3 connection
# ==============================================================================
s3_client = boto3.client(
    "s3",
    aws_access_key_id=key_id,
    aws_secret_access_key=secret_key,
    endpoint_url=endpoint,
    use_ssl=True
)

# Upload ttf and stress models
# ==============================================================================
s3_client.upload_file('models/battery_stress_model/saved_model.xml', bucket_name, Key='models/stress-detection/1/saved_model.xml')
s3_client.upload_file('models/battery_stress_model/saved_model.bin', bucket_name, Key='models/stress-detection/1/saved_model.bin')

s3_client.upload_file('models/battery_ttf_model/saved_model.xml', bucket_name, Key='models/time-to-failure/1/saved_model.xml')
s3_client.upload_file('models/battery_ttf_model/saved_model.bin', bucket_name, Key='models/time-to-failure/1/saved_model.bin')
