==> 08_query_models.ipynb <==
# Querying models

import requests
import json
import subprocess

# Query stress detection model
url = "https://stress-detection-ai-edge-project.apps.sno.pemlab.rdu2.redhat.com/v2/models/stress-detection/infer"

token = subprocess.check_output(["oc", "whoami", "-t"], text=True).strip()
headers = {
    "Authorization": f"Bearer {token}",
    "Content-Type": "application/json"
}

data = {
    "inputs": [
        {
            "name": "inputs",
            "shape": [1, 8], 
            "datatype": "FP32",
            "data": [[0.15, 97.00, 480.00, 355.00, 150.00, 50.00, 55.00, 22.00]]
        }
    ]
}

response = requests.post(url, headers=headers, json=data, verify=True)


def process_response(response):
    if response.status_code == 200:
        result = response.json()["outputs"][0]["data"][-1]
        print("Stress Prediction:", "STRESSED" if result == 1.0 else "NORMAL")
    else:
        print(f"Error: {response.status_code}, {response.text}")


process_response(response)

# Query time to failure model
url = "https://time-to-failure-ai-edge-project.apps.sno.pemlab.rdu2.redhat.com/v2/models/time-to-failure/infer"

token = subprocess.check_output(["oc", "whoami", "-t"], text=True).strip()
headers = {
    "Authorization": f"Bearer {token}",
    "Content-Type": "application/json"
}

data = {
    "inputs": [
        {
            "name": "inputs",
            "shape": [1, 5], 
            "datatype": "FP32",
            "data": [[45.0, 500.0, 370.0, 0.4, 98.0]]
        }
    ]
}

response = requests.post(url, headers=headers, json=data, verify=True)


def process_response(response):
    if response.status_code == 200:
        result = response.json()
        output_value = result["outputs"][0]["data"][0]  
        print(f"Predicted Time Before Failure: {output_value:.2f} hours")
    else:
        print(f"Error: {response.status_code}, {response.text}")


process_response(response)
